fix: explicit elytra handling suppression while in flight
